# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {"title": "Cridaa Booking API", "version": "1.0.0"},
  "paths": {
    "/api/health": {"get": {"summary": "Health check", "responses": {"200": {"description": "OK"}}}},
    "/api/auth/signup": {"post": {"summary": "Create account", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}, "email": {"type": "string", "format": "email"}}, "required": ["username", "password", "email"]}}}}, "responses": {"200": {"description": "Account created, token returned"}, "400": {"description": "Missing fields"}, "409": {"description": "User exists"}}}},
    "/api/auth/login": {"post": {"summary": "Login", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}}, "required": ["username", "password"]}}}}, "responses": {"200": {"description": "Token returned"}, "401": {"description": "Invalid credentials"}}}},
    "/api/slots": {"get": {"summary": "Available slots", "responses": {"200": {"description": "List of available slots"}}}},
    "/api/slots/{id}/book": {"post": {"summary": "Book slot", "security": [{"bearerAuth": []}], "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Booked"}, "401": {"description": "Unauthorized"}, "404": {"description": "Not found"}, "409": {"description": "Already booked"}}}},
    "/api/slots/{id}/cancel": {"delete": {"summary": "Cancel booking", "security": [{"bearerAuth": []}], "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Cancelled"}, "401": {"description": "Unauthorized"}, "403": {"description": "Not your booking"}, "404": {"description": "Not found"}}}},
    "/api/slots/mine": {"get": {"summary": "My bookings", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "List of user's bookings"}, "401": {"description": "Unauthorized"}}}}
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}